🚀 Feat/permission

+++ src/main/java/run/hxtia/workbd/service/admin/AdminUserService.java
@@ -65,5 +65,5 @@ public interface AdminUserService extends IService<AdminUsers> {
      * @param id：用户ID
      * @return ：用户信息
      */
-    AdminUserInfoDto getAdminUserInfo(Integer id);
+    AdminUserInfoDto getAdminUserInfo(Long id);
 }
